Stream .zip files to client with content-length included

+++ AnonymousPhotoBin/Data/FileMetadata.cs
@@ -25,7 +25,7 @@ namespace AnonymousPhotoBin.Data {
 
         public string Category { get; set; }
 
-        public long Size { get; set; }
+        public int Size { get; set; }
 
         [Column(TypeName = "binary(32)")]
         public byte[] Sha256 { get; set; }
